refactor(mail): 패키지 변경
