+++ mail-auth-client/build.gradle.kts
@@ -18,6 +18,8 @@ plugins {
 
 dependencies {
     implementation(project(":hongit-core"))
+
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     implementation("org.springframework.boot:spring-boot-starter-mail")
 
     testImplementation(project(":hongit-core"))
